# fix: Fix WeakMap set error, arg order irregularities