@keithamus 2016-11-14
This all looks great to me, but I'd like someone else like @meeber, @lucasfcosta, @vieiralucas or @shvaikalesh to review it also. Big 👍 from me though!