@meeber 2016-11-14
These seem like great changes. I experimented with performing the `isPrimitive` checks once at the top of `extensiveDeepEqual` instead of during every call to `memoizeCompare` and `memoizeSet` but didn't notice any gain in performance nor readability. Other than a minor nitpick about which `describe` block one of the new tests appears in, this looks good to go. Thanks @STRML!