@STRML 2016-11-14
Thanks for the review @meeber; I've moved that test and removed an unneeded `===` on L158.