@lucasfcosta 2016-11-14
LGTM too! Awesome job @STRML, thanks for your PR!